refactor: Move SerdeVec, PossiblyStr to dfx-core crate
