+++ src/dfx-core/Cargo.toml
@@ -18,6 +18,7 @@ k256 = { version = "0.11.4", features = ["pem"] }
 keyring = "1.2.0"
 lazy_static = "1.4.0"
 ring = "0.16.11"
+schemars = "0.8"
 sec1 = { version = "0.3.0", features = ["std"] }
 serde = "1.0"
 serde_json = "1.0.79"
