+++ src/dfx/src/util/mod.rs
@@ -4,7 +4,6 @@ use crate::{error_invalid_argument, error_invalid_data, error_unknown};
 use anyhow::Context;
 use candid::parser::typing::{pretty_check_file, TypeEnv};
 use candid::types::{Function, Type};
-use candid::Deserialize;
 use candid::{parser::value::IDLValue, IDLArgs};
 use fn_error_context::context;
 #[cfg(unix)]
@@ -12,13 +11,8 @@ use net2::unix::UnixTcpBuilderExt;
 use net2::{TcpBuilder, TcpListenerExt};
 use num_traits::FromPrimitive;
 use rust_decimal::Decimal;
-use schemars::JsonSchema;
-use serde::Serialize;
-use std::convert::TryFrom;
-use std::fmt::Display;
 use std::io::{stdin, Read};
 use std::net::{IpAddr, SocketAddr};
-use std::str::FromStr;
 use std::time::Duration;
 
 pub mod assets;
@@ -248,56 +242,6 @@ pub fn blob_from_arguments(
             Ok(typed_args)
         }
         v => Err(error_unknown!("Invalid type: {}", v)),
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, JsonSchema)]
-#[serde(untagged)]
-pub enum SerdeVec<T> {
-    One(T),
-    Many(Vec<T>),
-}
-
-impl<T> SerdeVec<T> {
-    pub fn into_vec(self) -> Vec<T> {
-        match self {
-            Self::One(t) => vec![t],
-            Self::Many(ts) => ts,
-        }
-    }
-}
-
-impl<T> Default for SerdeVec<T> {
-    fn default() -> Self {
-        Self::Many(vec![])
-    }
-}
-
-#[derive(Serialize, serde::Deserialize)]
-#[serde(untagged)]
-enum PossiblyStrInner<T> {
-    NotStr(T),
-    Str(String),
-}
-
-#[derive(Serialize, Deserialize, Default, Copy, Clone, Debug, JsonSchema)]
-#[serde(try_from = "PossiblyStrInner<T>")]
-pub struct PossiblyStr<T>(pub T)
-where
-    T: FromStr,
-    T::Err: Display;
-
-impl<T> TryFrom<PossiblyStrInner<T>> for PossiblyStr<T>
-where
-    T: FromStr,
-    T::Err: Display,
-{
-    type Error = T::Err;
-    fn try_from(inner: PossiblyStrInner<T>) -> Result<Self, Self::Error> {
-        match inner {
-            PossiblyStrInner::NotStr(t) => Ok(Self(t)),
-            PossiblyStrInner::Str(str) => T::from_str(&str).map(Self),
-        }
     }
 }
